debut d'initialisation avec setup du server ainsi que création du footer nav et carte principale
