+++ api/app.js
@@ -4,14 +4,15 @@ import dotenv from "dotenv";
 import pool from "./config.js";
 
 dotenv.config();
-
-app.use(cors(), {
-  origin: process.env.CLIENT_URL,
-  credentials: true,
-});
-app.use(express.json());
-
 const app = express();
+
+app.use(
+  cors({
+    origin: process.env.CLIENT_URL,
+    credentials: true,
+  })
+);
+app.use(express.json());
 
 async function connect() {
   await pool
