+++ client/src/component/card/Card.jsx
@@ -1,0 +1,19 @@
+import "./card.scss";
+
+export default function Card({
+  height,
+  backgroundHeader,
+  title,
+  subtitle,
+  content,
+}) {
+  return (
+    <article className="card" style={{ height: height }}>
+      <div className="cardHeader" style={{ backgroundColor: backgroundHeader }}>
+        <h2>{title}</h2>
+        <h3>{subtitle}</h3>
+      </div>
+      {content}
+    </article>
+  );
+}
