+++ client/src/component/card/card.scss
@@ -1,0 +1,20 @@
+.card {
+    position: relative;
+    border : 1px solid var(--color-border);
+    border-radius : var(--border-radius);
+    background-color: var(--color-secondary-light);
+    width: 90%;
+    padding: 1rem;
+
+    .cardHeader {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        border-top-left-radius: var(--border-radius);
+        border-top-right-radius: var(--border-radius)  ;
+        padding: 0.4rem;
+        padding-left: 2rem;
+        border-bottom: 1px solid var(--color-border);
+    }
+}
